Apply verified Mac UE 5.8.2 AshlineApple compile fixes.
Disable PCHs to avoid FVector vs CarbonCore, use PublicFrameworks for
GameController/CoreHaptics, keep native Apple hooks in a plain C header
and Unreal-free .mm, and cast weapon class for DualSense feedback.

+++ Source/Ashline/Weapons/AshlineWeaponComponent.cpp
@@ -147,7 +147,7 @@ void UAshlineWeaponComponent::FireShot()
 
 	if (UAshlineDualSense* DualSense = GEngine ? GEngine->GetEngineSubsystem<UAshlineDualSense>() : nullptr)
 	{
-		DualSense->TriggerWeaponFeedback(Active.Definition.Class, Active.Stats.RecoilPitch);
+		DualSense->TriggerWeaponFeedback(static_cast<uint8>(Active.Definition.Class), Active.Stats.RecoilPitch);
 	}
 
 	UE_LOG(LogAshline, Verbose, TEXT("Fired %s mag=%d reserve=%d"), *Active.Definition.WeaponId.ToString(), Active.AmmoInMag, Active.Reserve);
